Refactor importing content
Previously, we had the `src/content` folder that would source from the
build, and that file could be sourced anywhere built data was needed.
The problem with that is that for tests, we couldn't mock built data
which means we had to use built data in tests, which can be a problem.
Currently, some tests depended on the existence of at least one article,
which is clearly bad.

This commit changes it into the file `src/imports.js`, which sources
them from the build and puts them on a global variable `window.imports`
and is being imported by the `index.html` directly. On tests, we have to
source main by use Node's `import()` function, which also allows us to
source a different file that sets this global imports object with mocked
data.

Now most tests (unless those that are testing the build, such as the one
that checks if all article links are pointing to existing articles which
prevents typos from happening) can have mocked articles and other built
data for them, and also the project is now cleaner.

This commit also moves the `src/build` directory out of `src` and into a
new `scripts` directory, this way we can keep `src` with only code that
will go into the final app and with code that runs on the browser, while
we keep helper scripts on `scripts`, with code that will run on Node and
that can still be tested.
