+++ index.html
@@ -5,7 +5,8 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>Worldbuilding</title>
         <link href="assets/fonts/fonts.css" rel="stylesheet" />
-        <link href="build/output.css" rel="stylesheet" />
+        <link href="build/stylesheet.css" rel="stylesheet" />
+        <script type="module" src="src/imports.js"></script>
         <script type="module" src="src/main.js"></script>
     </head>
