+++ src/imports.js
@@ -1,0 +1,3 @@
+import data from "../build/articles.js";
+
+window.imports = { articles: data };
