+++ tests/mocks/imports.js
@@ -1,0 +1,3 @@
+import { data } from "./articles.js";
+
+window.imports = { articles: data };
